+++ tweaks/panel/src/SCIPanelRoot.m
@@ -18,7 +18,7 @@
 #import <objc/message.h>
 #import <objc/runtime.h>
 
-NSString *SCIVersionString = @"v0.9.34";  // AlbrhiPanel
+NSString *SCIVersionString = @"v0.9.35";  // AlbrhiPanel
 
 ///
 /// Albrhi's own control panel, in the iOS Settings app.